Add explanation texts after each choice
By default, the visitor will be shown an explanation text after each
choice is made. The choice description will be swapped with the
explanation, and the visitor will be prompted to advance to the next
question manually.

This may be optionally disabled for those who prefer not to be prompted.

Closes #93